chore: Separates builder into another module
