+++ src/lib.rs
@@ -7,7 +7,7 @@ pub(crate) mod utils;
 pub mod prelude {
     pub(crate) type GraphInteractionResult<T> = Result<T, GraphInteractionError>;
     pub use crate::directed::acyclic::DirectedAcyclicGraph;
+    pub use crate::directed::builder::DirectedGraphBuilder;
     pub use crate::directed::DirectedGraph;
-    pub use crate::directed::DirectedGraphBuilder;
     pub use crate::error::*;
 }
